Split shared lightwalletd test code into a module
